separate functions to a generator folder
