+++ packages/@aws-c2a/broadening-permissions/lib/generators/component.ts
@@ -1,0 +1,9 @@
+import { Component, Rule } from '@aws-c2a/rules';
+export { Rule } from '@aws-c2a/rules';
+
+export function generateComponent(resource: string, root: Rule): { component: Component, rule: Rule } {
+  const identifier = resource.replace(/::/g, '');
+  const component = Component.fromResource(identifier, resource);
+  const rule = root.createChild({ bindings: [component] });
+  return { component, rule };
+}
