+++ packages/@aws-c2a/broadening-permissions/lib/generators/index.ts
@@ -1,0 +1,4 @@
+export * from './component';
+export * from './high-risk-child';
+export * from './path';
+export * from './policy-statement';
