+++ packages/@aws-c2a/broadening-permissions/lib/security-changes.ts
@@ -1,5 +1,6 @@
 import { RuleRisk } from '@aws-c2a/models';
-import { CUserRule, CUserRules, Change, ChangeType, Component, ConditionOptions, Rule } from '@aws-c2a/rules';
+import { CUserRule, CUserRules, Change, ChangeType, Rule } from '@aws-c2a/rules';
+import { generateComponent, generateHighRiskChild, generateStatementRules } from './generators';
 import {
   IAM_INLINE_IDENTITY_POLICIES,
   IAM_INLINE_RESOURCE_POLICIES,
@@ -130,99 +131,3 @@ export class SecurityChangesRules {
   }
 }
 
-function generateComponent(resource: string, root: Rule): { component: Component, rule: Rule } {
-  const identifier = resource.replace(/::/g, '');
-  const component = Component.fromResource(identifier, resource);
-  const rule = root.createChild({ bindings: [component] });
-  return { component, rule };
-}
-
-/**
- * Generates a string array to represent the path to the PolicyDocument for statements.
- *
- * For example the path to a PolicyDocument for inline identity policies are the following:
- * ```json
- * // AWS::IAM::User
- * "Properties": {
- *   "Policies": [{
- *     "PolicyDocument": {
- *       "Statement": [
- *         {...}
- *       ]
- *     }
- *   }]
- * }
- * ```
- * The query for this path would be: `this._generateStatementPath('PolicyDocument', 'Policies', '*')`
- *
- * Some policies call the PolicyDocument by another name:
- * ```json
- * // AWS::IAM::Policy
- * "Properties": {
- *   "AssumeRolePolicyDocument": {
- *     "Statement": [
- *       {...}
- *     ]
- *   }
- * }
- * ```
- * The query for this path would be: `this._generateStatementPath('AssumeRolePolicyDocument')`
- *
- * @param documentName the key for the policy document [default: PolicyDocument]
- * @param propertyPrefix the strings that prefix the policy document
- *
- * @returns a property path with the form
- * ['Properties', ...propertyPrefix, documentName, 'Statement', '*']
- */
-function generateStatementPath(documentName = 'PolicyDocument', ...propertyPrefix: string[]): string[] {
-  return ['Properties', ...propertyPrefix, documentName, 'Statement', '*'];
-}
-
-/**
- * See {@link generateStatementPath}.
- *
- * @param documentName the key for the policy document [default: PolicyDocument]
- * @param propertyPrefix the strings that prefix the policy document
- *
- * @returns a property path with the form
- * ['Properties', ...propertyPrefix, documentName, 'Statement', '*', 'Effect']
- */
-function generateEffectPath(documentName = 'PolicyDocument', ...propertyPrefix: string[]): string [] {
-  return [...generateStatementPath(documentName, ...propertyPrefix), 'Effect'];
-}
-
-function generateStatementRules(parent: Rule, component: Component, documentName = 'PolicyDocument', ...prefix: string[]) {
-  generateHighRiskChild(parent, component, {
-    change: Change.INSERT,
-    equals: 'Allow',
-    sourcePath: generateEffectPath(documentName, ...prefix),
-  });
-  generateHighRiskChild(parent, component, {
-    change: Change.INSERT_PROP,
-    equals: 'Allow',
-    targetPath: generateStatementPath(documentName, ...prefix),
-    sourcePath: ['Effect'],
-  });
-  generateHighRiskChild(parent, component, {
-    change: Change.UPDATE_PROP,
-    equals: 'Allow',
-    targetPath: generateStatementPath(documentName, ...prefix),
-    sourcePath: ['Effect'],
-  });
-}
-
-interface HighRiskChildOptions extends ConditionOptions {
-  change: Change;
-  equals?: string;
-}
-
-function generateHighRiskChild(parent: Rule, component: Component, {change, ...options}: HighRiskChildOptions): void {
-  parent.createChild({
-    conditions: [
-      change.appliesTo(component, { targetPath: options?.targetPath }),
-      ...(options.equals ? [change.equals(options.equals, ChangeType.NEW, { sourcePath: options?.sourcePath })] : []),
-    ],
-    risk: RuleRisk.High,
-    target: change,
-  });
-}
